#74 remove empty views and routes, add pass-eye in login form, rename user to profile
