+++ front/src/components/mmt-header/mmt-header.component.html
@@ -34,7 +34,7 @@
       <v-btn icon class="white--text">
         <v-icon>mdi-magnify</v-icon>
       </v-btn>
-    
+
     </v-app-bar>
   </v-container>
