+++ front/src/components/mmt-header/mmt-header.component.js
@@ -19,11 +19,8 @@ export default {
     links () {
       return [
         { title: 'menu.accommodation', route: '/accommodation', isVisible: true },
-        { title: 'menu.transport', route: '/transport', isVisible: true },
         { title: 'menu.leisure', route: '/leisure', isVisible: true },
         { title: 'menu.food', route: '/food', isVisible: true },
-        { title: 'menu.hotTours', route: '/hot-tours', isVisible: true },
-        { title: 'menu.articles', route: '/articles', isVisible: true },
         { title: 'menu.signin', route: '/login', isVisible: !this.isAuthorize },
         { title: 'menu.signup', route: '/signup', isVisible: !this.isAuthorize }
       ];
