+++ front/src/locales/translations/enUS.json
@@ -3,11 +3,8 @@
   "menu": {
     "aboutUs": "About Us",
     "accommodation": "Accommodation",
-    "transport": "Transport",
     "leisure": "Lleisure",
     "food": "Food",
-    "hotTours": "Hot Tours",
-    "articles": "Articles",
     "privacyPolicy": "Privacy policy",
     "cookiePolicy": "Cookie Policy",
     "termsOfUse": "Terms of use",
@@ -16,13 +13,11 @@
     "reviewNew": "Leave a review about the site",
     "tickets": "Tickets",
     "travelRoute": "Travel Route",
-    "bookmarks": "Bookmarks",
     "scheduler": "Scheduler",
     "history": "History",
     "historyPrev": "Previous",
     "historyNext": "Upcoming",
     "settings": "Settings",
-    "notifications": "Notifications",
     "editProfile": "Edit Profile",
     "signin": "Login",
     "signup": "Register",
