+++ front/src/locales/translations/ru.json
@@ -3,11 +3,8 @@
   "menu": {
     "aboutUs": "О нас",
     "accommodation": "Жилье",
-    "transport": "Транспорт",
     "leisure": "Досуг",
     "food": "Еда",
-    "hotTours": "Горящие туры",
-    "articles": "Статьи",
     "privacyPolicy": "Политика конфиденциальности",
     "cookiePolicy": "Политика использования Cookie",
     "termsOfUse": "Пользовательское соглашение",
@@ -16,13 +13,11 @@
     "reviewNew": "Оставить отзыв о сайте",
     "tickets": "Билеты",
     "travelRoute": "Маршрут",
-    "bookmarks": "Закладки",
     "scheduler": "Планер",
     "history": "История",
     "historyPrev": "Предыдущие",
     "historyNext": "Предстоящие",
     "settings": "Настройки",
-    "notifications": "Уведомления",
     "editProfile": "Редактировать профиль",
     "signin": "Вход",
     "signup": "Регистрация",
