+++ front/src/locales/translations/uk.json
@@ -3,11 +3,8 @@
   "menu": {
     "aboutUs": "Про нас",
     "accommodation": "Житло",
-    "transport": "Транспорт",
     "leisure": "Дозвілля",
     "food": "Їжа",
-    "hotTours": "Гарячі тури",
-    "articles": "Статті",
     "privacyPolicy": "Політика конфіденційності",
     "cookiePolicy": "Політика використання Cookie",
     "termsOfUse": "Угода",
@@ -16,13 +13,11 @@
     "reviewNew": "Залишити відгук про сайт",
     "tickets": "Квитки",
     "travelRoute": "Маршрут",
-    "bookmarks": "Закладки",
     "scheduler": "Планер",
     "history": "Історія",
     "historyPrev": "Попередні",
     "historyNext": "Майбутні",
     "settings": "Налаштування",
-    "notifications": "Повідомлення",
     "editProfile": "Редагувати профіль",
     "signin": "Вхід",
     "signup": "Реєстрація",
