--- front/src/views/Articles.vue
@@ -1,9 +1,0 @@
-<template>
-  <h2>Articles</h2>
-</template>
-
-<script>
-
-export default {
-};
-</script>
