--- front/src/views/HotTours.vue
@@ -1,9 +1,0 @@
-<template>
-  <h2>Hot Tours</h2>
-</template>
-
-<script>
-
-export default {
-};
-</script>
