--- front/src/views/Transport.vue
@@ -1,9 +1,0 @@
-<template>
-  <h2>Transport</h2>
-</template>
-
-<script>
-
-export default {
-};
-</script>
