+++ front/src/views/profile/Accommodation.vue
