+++ front/src/views/profile/entertainments-new/entertainments-new.view.js
