+++ front/src/views/profile/entertainments/entertainments.view.js
