+++ front/src/views/profile/entertainments/entertainments.view.vue
