+++ front/src/views/sign-in/sign-in.view.js
@@ -8,6 +8,7 @@ export default {
   },
   data () {
     return {
+      showFirstPassword: false,
       email: '',
       password: '',
       serverError: '',
