--- front/src/views/user/Food.vue
@@ -1,16 +1,0 @@
-<template>
-  <div>
-    <mmt-stepper></mmt-stepper>
-  </div>
-</template>
-
-<script>
-// @ is an alias to /src
-import { MmtStepper } from '@components';
-
-export default {
-  components: {
-    MmtStepper
-  }
-};
-</script>
